DABOM-229 refactor: UsageSyncServiceImpl 책임 분리
